+++ src/clj/rems/api/services/attachment.clj
@@ -1,13 +1,16 @@
 (ns rems.api.services.attachment
   (:require [clojure.set :as set]
-            [clojure.test :refer :all]
+            [clojure.tools.logging :as log]
             [rems.application.commands :as commands]
             [rems.common.application-util :as application-util]
             [rems.auth.util :refer [throw-forbidden]]
             [rems.db.applications :as applications]
             [rems.db.attachments :as attachments]
-            [ring.util.http-response :refer [ok content-type header]])
-  (:import [java.io ByteArrayInputStream]))
+            [rems.util :refer [getx]]
+            [ring.util.http-response :refer [ok content-type header]]
+            [ring.util.io :as ring-io])
+  (:import [java.io ByteArrayInputStream ByteArrayOutputStream]
+           [java.util.zip ZipOutputStream ZipEntry ZipException]))
 
 (defn download [attachment]
   (-> (ok (ByteArrayInputStream. (:attachment/data attachment)))
@@ -41,3 +44,23 @@
                     (:application/permissions application))
       (throw-forbidden))
     (attachments/save-attachment! file user-id application-id)))
+
+(defn zip-attachments [application]
+  (let [zip-input (ring-io/piped-input-stream
+                   (fn [out]
+                     (with-open [zip (ZipOutputStream. out)]
+                       (doseq [metadata (getx application :application/attachments)]
+                         (let [id (getx metadata :attachment/id)
+                               attachment (attachments/get-attachment id)]
+                           ;; we deduplicate filenames when uploading, but here's a
+                           ;; failsafe in case we have duplicate filenames in old
+                           ;; applications
+                           (try
+                             (.putNextEntry zip (ZipEntry. (getx attachment :attachment/filename)))
+                             (.write zip (getx attachment :attachment/data))
+                             (.closeEntry zip)
+                             (catch ZipException e
+                               (log/warn "Ignoring attachment" (pr-str metadata) "when generating zip. Cause:" e))))))))]
+    (-> (ok zip-input)
+        (header "Content-Disposition" (str "attachment;filename=attachments-" (getx application :application/id) ".zip"))
+        (content-type "application/zip"))))